src/requirements.txt
matplotlib==3.5.2
numpy==1.21.6
opencv-python==4.6.0.66
Pillow==9.3.0
tqdm==4.64.0
tensorboard==2.9.1
torch==1.12.0
torchvision==0.13.0
scikit-learn==1.0.2
kornia==0.5.7
scikit-image==0.19.3
deap==1.3.1
ftfy==6.1.1
regex==2022.7.25
pandas==1.5.0